[Refactor] refactoring for mDPP
